% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections_two_three_circles.R
\name{intersection_two_circles}
\alias{intersection_two_circles}
\alias{intersection_three_circles}
\title{Auxiliary functions for computing circle intersection areas}
\usage{
intersection_two_circles(centers_x, centers_y, radii)

intersection_three_circles(centers_x, centers_y, radii)
}
\arguments{
\item{centers_x, centers_y, radii}{Numeric vectors of length N with the
coordinates of the center of the circles in x and y, and the radius
respectively.}
}
\description{
`intersection_two_circles()` Returns the area of intersection of two circles.

 `intersection_three_circles()` Returns the area of intersection of three circles.
}
