% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circles_simple.R
\name{plot_circles_simple}
\alias{plot_circles_simple}
\title{Basic plot of circles}
\usage{
plot_circles_simple(centers_x, centers_y, radii, npoints = 500)
}
\arguments{
\item{centers_x, centers_y, radii}{Numeric vectors of length N with the
coordinates of the center of the circles in x and y, and the radius
respectively.}

\item{npoints}{The number of points to plot each circle with. Defaults to 500}
}
\description{
Generates a very simple diagram of circles, useful
for testing the functions of this package. The text in the plot indicates the
area of each circle. The number of each circle is written on top of itself.
}
\examples{
# Example of intersection areas including a Reuleaux triangle
x <- c(0, 1, 0.5)
y <-c(0, 0, sqrt(1-0.5**2))
radii <- c(1, 1, 1)
plot_circles_simple(centers_x = x, centers_y = y, radii = radii)

# Example with more circles
x2 <- c(0, 4, 2, 4, 5)
y2 <- c(1, 5, 4, 2, 1)
radii2 <- c(1, 4 ,2, 2, 1)
plot_circles_simple(centers_x = x2, centers_y = y2, radii = radii2)
}
\author{
Hugo Salinas \email{hugosal@comunidad.unam.mx}.
}
